Update retrospective.md
I have made some contributions to the retrospective.md file. It was a succinct overview of my personal experience during this project.

Additionally, I observed that I can not edit this file directly. Consequently, I have created a branch for this commit and will start a pull request. I humbly appeal that my commit be promptly reviewed and added to the main branch.